Make image return inline when requested from hipchat
Previously, the response from the orly instruction would return a link.
While this is all well and good, people had to click things to see it.
This change simply appends a &.png suffix on the link returned, which
means that hipchat will render the image right in the chat window.

Signed-off-by: Michael Stahnke <stahnma@puppet.com>